==> 2-surg.ipynb <==
import pandas as pd

info = pd.read_excel("data/info.xlsx")

surg = pd.read_csv("data/patient_procedure events.csv", dtype=str)

surg = surg[['MRN','LOG_ID','EVENT_DISPLAY_NAME','EVENT_TIME','NOTE_TEXT']]
surg['MRN'] = surg['MRN'].astype('string')
surg['LOG_ID'] = surg['LOG_ID'].astype('string')
surg['EVENT_DISPLAY_NAME'] = surg['EVENT_DISPLAY_NAME'].astype('string')
surg['EVENT_TIME'] = pd.to_datetime(surg['EVENT_TIME'], format="%m/%d/%y %H:%M")
surg['NOTE_TEXT'] = surg['NOTE_TEXT'].astype('string')

surg = surg.sort_values(by=['MRN', 'EVENT_TIME', 'LOG_ID'])

print("number of duplicate rows: ", surg[surg.duplicated()].shape)
surg = surg.drop_duplicates()
surg = surg.dropna(how='any') #axis{0 or ‘index’, 1 or ‘columns’}, default 0

# surg = surg.loc[surg['EVENT_DISPLAY_NAME'].eq('Anesthesia Ready') | surg['EVENT_DISPLAY_NAME'].eq('Emergence')]
surg = surg.loc[surg['EVENT_DISPLAY_NAME'].eq('Anesthesia Ready')]
surg = surg.loc[surg['MRN'].isin(info['MRN']) & surg['LOG_ID'].isin(info['LOG_ID'])]

surg.to_excel('data/surg.xlsx', index=False)
