==> 3-events.ipynb <==
import pandas as pd

info = pd.read_excel("data/info.xlsx")
surg = pd.read_excel("data/surg.xlsx")

**CREATE EVENT LOG**

admsn = pd.DataFrame({'patient':info.MRN, 'log':info.LOG_ID, 'action':'Hospital Admission', 'datetime':info['HOSP_ADMSN_TIME']})
disch = pd.DataFrame({'patient':info.MRN, 'log':info.LOG_ID, 'action':'Hospital Discharge', 'datetime':info['HOSP_DISCH_TIME']})
or_in = pd.DataFrame({'patient':info.MRN, 'log':info.LOG_ID, 'action':'Operating Room Enter', 'datetime':info['IN_OR_DTTM']})
or_out = pd.DataFrame({'patient':info.MRN, 'log':info.LOG_ID, 'action':'Operating Room Exit', 'datetime':info['OUT_OR_DTTM']})
an_start = pd.DataFrame({'patient':info.MRN, 'log':info.LOG_ID, 'action':'Anesthesia Start', 'datetime':info['AN_START_DATETIME']})
an_stop = pd.DataFrame({'patient':info.MRN, 'log':info.LOG_ID, 'action':'Anesthesia Stop', 'datetime':info['AN_STOP_DATETIME']})
planned = pd.DataFrame({'patient':info.MRN, 'log':info.LOG_ID, 'action':'Surgery Planned', 'datetime':info['SRG_PLN_TIME']})
cancelled_rows = info[info.SURGERY_TYPE == 'Surgery Cancelled']
cancelled = pd.DataFrame({'patient':cancelled_rows.MRN, 'log':cancelled_rows.LOG_ID, 'action':'Surgery Cancelled', 'datetime':cancelled_rows['SRG_CNL_TIME']})
passed_rows =  info[info.SURGERY_TYPE == 'Surgery Date Passed']
passed = pd.DataFrame({'patient':passed_rows.MRN, 'log':passed_rows.LOG_ID, 'action':'Surgery Passed', 'datetime':passed_rows['SRG_PLN_TIME']})
# find elements in planned that are not in passed
planned = planned[~(planned['patient'].isin(passed['patient']) & planned['log'].isin(passed['log']))].reset_index(drop=True)

sg_start_rows = surg[surg.EVENT_DISPLAY_NAME == 'Anesthesia Ready']
sg_start = pd.DataFrame({'patient':sg_start_rows.MRN, 'log':sg_start_rows.LOG_ID, 'action':'Anesthesia Ready', 'datetime':sg_start_rows['EVENT_TIME']})
sg_stop_rows = surg[surg.EVENT_DISPLAY_NAME == 'Emergence']
sg_stop = pd.DataFrame({'patient':sg_stop_rows.MRN, 'log':sg_stop_rows.LOG_ID, 'action':'Surgery Stop', 'datetime':sg_stop_rows['EVENT_TIME']})

events = pd.concat([admsn, disch, or_in, or_out, an_start, an_stop, planned, cancelled, passed, sg_start, sg_stop])
events = events[['patient', 'log', 'action', 'datetime']]
events['patient'] = events['patient'].astype("string")
events['log'] = events['log'].astype("string")
events['action'] = events['action'].astype("string")
events['datetime'] = pd.to_datetime(events['datetime'], format="%m/%d/%y %H:%M")
events = events.drop_duplicates(subset=['patient', 'action', 'datetime'])

# Define custom sorting order for events
custom_order = {
    'Surgery Passed': 1,
    'Hospital Admission': 2,
    'Surgery Planned': 3,
    'Surgery Cancelled': 4,
    "Operating Room Enter": 5,
    "Anesthesia Start": 6,
    'Anesthesia Ready':7,
    # 'Surgery Stop':8,
    "Anesthesia Stop": 8,
    "Operating Room Exit": 9,
    'Hospital Discharge':10
}

events = events.sort_values(
    by=['patient', 'datetime', 'log'],
).sort_values(
    by=['patient', 'datetime', 'log'],
    key=lambda col: col.map(custom_order).fillna(0),  # Default 0 for other events
    kind= "stable"
)

events = events.dropna(axis=0)
events.reset_index(drop=True, inplace=True)

events.to_excel("data/events.xlsx", index=False)
