==> 4-log.ipynb <==
import pandas as pd
import pm4py

events = pd.read_excel("data/events.xlsx")

log = pm4py.format_dataframe(events, case_id='patient', activity_key='action', timestamp_key='datetime', timest_format='%Y-%m-%d %H:%M:%S')
log = log[['case:concept:name', 'log', 'concept:name', 'time:timestamp', '@@index', '@@case_index']]
log['time:timestamp'] = log['time:timestamp'].dt.tz_localize(None)

log.to_excel("data/log.xlsx", index=False)
