==> 6-conformance.ipynb <==
import pandas as pd
import pm4py

log = pd.read_excel("data/log.xlsx")

# Step 1: Discover process model using Inductive Miner
petri_net, initial_marking, final_marking = pm4py.discover_petri_net_inductive(log)

# Step 2: Perform conformance checking with token-based replay
replay_result = pm4py.conformance_diagnostics_token_based_replay(log, petri_net, initial_marking, final_marking)

# Check the conformance statistics
print("Conformance checking (token-based replay):")
for i in replay_result:
    print(i)

# we compare the original log versus the discovered model
fitness = pm4py.fitness_token_based_replay(log, petri_net, initial_marking, final_marking)

print("Fitness Values:")
print(fitness)

# we compare the original log versus the discovered model
precision = pm4py.precision_token_based_replay(log, petri_net, initial_marking, final_marking)
print(precision) # 0.4304572926352077

################## BE CAREFULL ' it is taking too long
# we check if the discovered model is a sound workflow net
#### is_sound_wfnet = pm4py.check_soundness(petri_net, initial_marking, final_marking)[0]
#### print(is_sound_wfnet) # True
