==> 8-transitions.ipynb <==
import pandas as pd

log = pd.read_excel("data/log.xlsx")

from pm4py.statistics.eventually_follows.pandas import get as eventually_follows
follow_df = eventually_follows.get_partial_order_dataframe(
    log,
    sort_caseid_required=False,
    sort_timestamp_along_case_id=False,
    reduce_dataframe=False,
)

# Create transition and flow time in minutes columns
follow_df["transition"] = follow_df["concept:name"] + " > " + follow_df["concept:name_2"]
follow_df['@@flow_time_minute'] = follow_df['@@flow_time'] / 60

date_columns = follow_df.select_dtypes(include=['datetime64[ns, UTC]']).columns
for date_column in date_columns:
    follow_df[date_column] = follow_df[date_column].dt.tz_localize(None)

follow_df.to_excel("data/transitions.xlsx", index=False)
